backup/temp/scripts/test_openbb_code.py
# -*- coding: utf-8 -*-
"""
===================================
OpenBB Code Availability Test
===================================

[Purpose]
Verify OpenBB code availability from GitHub backup

[Usage]
python scripts/test_openbb_code.py
"""

import os
import sys
from pathlib import Path

sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))


def test_openbb_import():
    """Test OpenBB import from installed package"""
    print("\n" + "=" * 60)
    print("Test 1: OpenBB Package Import")
    print("=" * 60)
    
    try:
        from openbb import obb
        print("SUCCESS: OpenBB package imported successfully")
        print(f"OpenBB version: {obb.version if hasattr(obb, 'version') else 'N/A'}")
        return True
    except ImportError as e:
        print(f"FAILED: OpenBB import failed: {e}")
        return False


def test_github_backup():
    """Test OpenBB GitHub backup availability"""
    print("\n" + "=" * 60)
    print("Test 2: GitHub Backup Availability")
    print("=" * 60)
    
    project_root = Path(__file__).parent.parent
    backup_path = project_root / "openbb_backup"
    
    if not backup_path.exists():
        print(f"FAILED: GitHub backup not found at {backup_path}")
        return False
    
    print(f"SUCCESS: GitHub backup found at {backup_path}")
    
    key_files = [
        "openbb_terminal",
        "README.md",
        "requirements.txt"
    ]
    
    for key_file in key_files:
        file_path = backup_path / key_file
        if file_path.exists():
            print(f"  - {key_file}: EXISTS")
        else:
            print(f"  - {key_file}: MISSING")
    
    return True


def test_dependencies():
    """Test required dependencies"""
    print("\n" + "=" * 60)
    print("Test 3: Required Dependencies")
    print("=" * 60)
    
    required_packages = [
        "openbb",
        "pandas",
        "numpy",
        "tushare",
        "akshare"
    ]
    
    all_ok = True
    for package in required_packages:
        try:
            __import__(package)
            print(f"  - {package}: INSTALLED")
        except ImportError:
            print(f"  - {package}: MISSING")
            all_ok = False
    
    return all_ok


def test_openbb_functionality():
    """Test basic OpenBB functionality"""
    print("\n" + "=" * 60)
    print("Test 4: Basic OpenBB Functionality")
    print("=" * 60)
    
    try:
        from openbb import obb
        
        print("Testing OpenBB equity price fetch...")
        
        try:
            df = obb.equity.price.historical("AAPL", provider="yfinance", days=5)
            if df is not None and not df.empty:
                print(f"SUCCESS: Fetched {len(df)} rows of AAPL data")
                return True
            else:
                print("WARNING: No data returned, but API is accessible")
                return True
        except Exception as e:
            print(f"WARNING: API call failed: {e}")
            print("This may be due to network issues, but OpenBB is installed correctly")
            return True
            
    except Exception as e:
        print(f"FAILED: OpenBB functionality test failed: {e}")
        return False


def main():
    """Run all tests"""
    print("\n" + "=" * 60)
    print("OpenBB Code Availability Test")
    print("=" * 60)
    
    results = {
        "OpenBB Import": test_openbb_import(),
        "GitHub Backup": test_github_backup(),
        "Dependencies": test_dependencies(),
        "Functionality": test_openbb_functionality()
    }
    
    print("\n" + "=" * 60)
    print("Test Summary")
    print("=" * 60)
    
    for test_name, result in results.items():
        status = "PASS" if result else "FAIL"
        print(f"  {test_name}: {status}")
    
    all_passed = all(results.values())
    
    print("\n" + "=" * 60)
    if all_passed:
        print("OVERALL: ALL TESTS PASSED")
    else:
        print("OVERALL: SOME TESTS FAILED")
    print("=" * 60)
    
    return all_passed


if __name__ == '__main__':
    success = main()
    sys.exit(0 if success else 1)
